Denormalize arrays, ids, and arrays of ids
This commit updates the denormalize function to handle denormalizing lists
when the schema is an IterableSchema (i.e. arrayOf(entitySchema)). I did some
refactoring as part of that work, and a nice side effect was being able to
denormalize just by passing the id (or array of ids). This was already how
denormalizing nested arrays was handled, so I think it makes sense.

Another cool side effect is the greater symmetry with normalizr. This is
now true:

```js
const { entities, result } = normalize(response, schema);

denormalize(result, entities, schema) === response
```